Extract `search_and_download` into `data.fetch` module
