+++ tests/integration/test_e2e.py
@@ -10,44 +10,12 @@ this library.
 
 from __future__ import annotations
 
-from pathlib import Path
-from typing import Literal
-
-import earthaccess
 import pandas as pd
 
 from iceflow.data.atm1b import atm1b_data
+from iceflow.data.fetch import search_and_download
 from iceflow.data.models import IceflowDataFrame
 from iceflow.itrf.converter import transform_itrf
-
-ShortName = Literal["ILATM1B"]
-
-
-# TODO: move this
-def search_and_download(
-    *,
-    version: str,
-    short_name: ShortName,
-    bounding_box,
-    temporal: tuple[str, str],
-    output_dir: Path,
-) -> list[Path]:
-    earthaccess.login()
-
-    results = earthaccess.search_data(
-        short_name=short_name,
-        version=version,
-        bounding_box=bounding_box,
-        temporal=temporal,
-    )
-
-    # short_name based subdir for data.
-    output_subdir = output_dir / short_name
-    output_subdir.mkdir(exist_ok=True)
-    downloaded_files = earthaccess.download(results, str(output_subdir))
-    downloaded_filepaths = [Path(filepath_str) for filepath_str in downloaded_files]
-
-    return downloaded_filepaths
 
 
 def test_e2e(tmp_path):
